refactor : 서비스의 역할을 세분화하여 예약, 공연, 알람에 관한 기능으로 분리함
- 파사드 패턴을 활용함
