+++ src/main/java/com/wanted/preonboarding/ticketing/service/PerformanceService.java
@@ -7,14 +7,12 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
 public class PerformanceService {
     private final PerformanceRepository performanceRepository;
 
-    @Transactional(readOnly = true)
     public Page<ReadPerformanceResponse> readPerformance(String isReserve, Pageable pageable) {
         Page<Performance> performances = performanceRepository.findByIsReserve(isReserve, pageable);
