+++ src/main/java/com/wanted/preonboarding/ticketing/service/ReservationService.java
@@ -7,11 +7,9 @@ import com.wanted.preonboarding.ticketing.domain.dto.request.CreateAlarmRequest;
 import com.wanted.preonboarding.ticketing.domain.dto.request.CreateReservationRequest;
 import com.wanted.preonboarding.ticketing.domain.dto.request.ReadReservationRequest;
 import com.wanted.preonboarding.ticketing.domain.dto.response.*;
-import com.wanted.preonboarding.ticketing.domain.entity.Alarm;
 import com.wanted.preonboarding.ticketing.domain.entity.Performance;
 import com.wanted.preonboarding.ticketing.domain.entity.PerformanceSeatInfo;
 import com.wanted.preonboarding.ticketing.domain.entity.Reservation;
-import com.wanted.preonboarding.ticketing.repository.AlarmRepository;
 import com.wanted.preonboarding.ticketing.repository.PerformanceRepository;
 import com.wanted.preonboarding.ticketing.repository.PerformanceSeatInfoRepository;
 import com.wanted.preonboarding.ticketing.repository.ReservationRepository;
@@ -29,9 +27,9 @@ public class ReservationService {
     private final ReservationRepository reservationRepository;
     private final PerformanceRepository performanceRepository;
     private final PerformanceSeatInfoRepository performanceSeatInfoRepository;
-    private final AlarmRepository alarmRepository;
 
-    private final AlarmSender alarmSender;
+    private final AlarmService alarmService;
+    private final PerformanceService performanceService;
     private final ReservationValidator reservationValidator;
 
     @Transactional
@@ -70,18 +68,16 @@ public class ReservationService {
         return reservations.map(Reservation::toReadReservationResponse);
     }
 
+    @Transactional(readOnly = true)
+    public Page<ReadPerformanceResponse> readPerformance(String isReserve, Pageable pageable) {
+        return performanceService.readPerformance(isReserve, pageable);
+    }
+
     @Transactional
     public CreateAlarmResponse createAlarm(CreateAlarmRequest createAlarmRequest) {
         Performance performance = performanceRepository.getReferenceById(createAlarmRequest.getPerformanceId());
-        Alarm alarm = saveAlarm(createAlarmRequest, performance);
 
-        return alarm.toCreateAlarmResponse();
-    }
-
-    private Alarm saveAlarm(CreateAlarmRequest createAlarmRequest, Performance performance) {
-        Alarm alarm = createAlarmRequest.from(performance);
-
-        return alarmRepository.save(alarm);
+        return alarmService.createAlarm(createAlarmRequest, performance);
     }
 
     @Transactional
@@ -89,7 +85,7 @@ public class ReservationService {
         deleteReservation(cancelReservationRequest);
         PerformanceSeatInfo performanceSeatInfo = changeSeatInfo(cancelReservationRequest);
 
-        return alarmSender.sendAlarm(performanceSeatInfo);
+        return alarmService.sendAlarm(performanceSeatInfo);
     }
 
     private PerformanceSeatInfo changeSeatInfo(CancelReservationRequest cancelReservationRequest) {
